Gjør LinkList stable
