--- packages/jokul/src/components-beta/link-list/index.ts
@@ -1,7 +1,0 @@
-export { LinkList } from "./LinkList.js";
-export { LinkListLink } from "./LinkListLink.js";
-
-export type {
-    LinkListProps,
-    LinkListLinkComponent,
-} from "./types.js";
