--- packages/jokul/src/components-beta/link-list/styles/_index.scss
@@ -1,2 +1,0 @@
-@forward "link-list";
-@use "../../../components/icon/styles";
