+++ packages/jokul/src/components/link-list/development/Example.tsx
@@ -1,16 +1,7 @@
 import React from "react";
 import { DevExample } from "utils/dev-example/DevExample.js";
-import {
-    OrderedLinkListExample,
-    UnorderedLinkListExample,
-    knobs,
-} from "./LinkListExample.js";
+import { LinkListExample } from "./LinkListExample.js";
 
 export default function Example() {
-    return (
-        <>
-            <DevExample component={OrderedLinkListExample} knobs={knobs} />
-            <DevExample component={UnorderedLinkListExample} knobs={knobs} />
-        </>
-    );
+    return <DevExample component={LinkListExample} />;
 }
