+++ packages/jokul/src/components/link-list/index.ts
@@ -1,12 +1,4 @@
 export { LinkList } from "./LinkList.js";
+export { LinkListLink } from "./LinkListLink.js";
 
-export type {
-    LinkListProps,
-    UnorderedLinkListProps,
-    OrderedLinkListProps,
-    LinkComponent,
-    ItemProps,
-} from "./types.js";
-
-export { LinkList as BETA_LinkList } from "../../components-beta/link-list/LinkList.js";
-export type { LinkListProps as BETA_LinkListProps } from "../../components-beta/link-list/types.js";
+export type { LinkListProps, LinkListLinkComponent } from "./types.js";
