+++ packages/jokul/src/components/link-list/styles/link-list.scss
@@ -1,88 +1,77 @@
 @use "../../../core/jkl";
+@use "../../icon/styles/base-styles" as icon;
 
 @layer jokul.components {
     .jkl-link-list {
-        --border-color: var(--jkl-color-border-separator);
+        --border: 1px solid var(--jkl-color-border-separator);
+        --text-color: var(--jkl-color-text-default);
 
-        display: flex;
-        flex-direction: column;
-        width: 100%;
-        padding: 0;
-        margin: 0;
-        border-bottom: jkl.rem(1px) solid var(--border-color);
-        height: min-content;
-        counter-reset: link-list;
+        @include jkl.text-style("paragraph-medium");
 
-        @media (hover: hover) {
-            &:has(:focus-visible),
-            &:has(:hover) {
-                .jkl-link-list-link:not(:hover):not(:focus-visible) {
-                    --text-color: var(--jkl-color-border-separator);
+        .jkl-link-list-title {
+            margin-block-end: var(--jkl-unit-10);
+
+            @include jkl.text-style("paragraph-small");
+        }
+
+        & ul {
+            display: flex;
+            flex-direction: column;
+            width: 100%;
+            padding: 0;
+            margin: 0;
+            height: min-content;
+            border-radius: var(--jkl-unit-05);
+            overflow: hidden;
+            background: var(--jkl-color-background-container-low);
+
+            .jkl-link-list-link {
+                display: flex;
+                align-items: center;
+                justify-content: space-between;
+                gap: var(--jkl-unit-20);
+                width: 100%;
+                text-decoration: none;
+                color: var(--text-color);
+                text-wrap: balance;
+                box-sizing: border-box;
+                padding: var(--jkl-unit-20);
+
+                @include jkl.text-style("text-medium");
+                @include jkl.motion("standard", "productive");
+                transition-property: background;
+
+                &::after {
+                    content: "arrow_forward" / "";
+
+                    @include icon.base-styles;
+                    @include jkl.motion("standard", "productive");
+                    transition-property: translate;
+                }
+
+                &:is(:hover, :focus-visible) {
+                    background: var(--jkl-color-background-interactive-selected);
+
+                    &::after {
+                        translate: 4px 0;
+                    }
                 }
             }
         }
-    }
 
-    .jkl-link-list-item {
-        display: flex;
-        counter-increment: link-list;
-    }
+        & .jkl-link-list-item {
+            display: flex;
 
-    .jkl-link-list-link {
-        --jkl-link-list-padding: var(--jkl-unit-10);
-        --text-color: var(--jkl-color-text-default);
-        --icon-background: transparent;
-
-        display: flex;
-        justify-content: space-between;
-        align-items: first baseline;
-        gap: var(--jkl-unit-20);
-        padding: var(--jkl-link-list-padding) 0;
-        border-top: jkl.rem(1px) solid var(--border-color);
-        width: 100%;
-        text-decoration: none;
-        color: var(--text-color);
-        cursor: pointer;
-
-        @include jkl.motion("standard", "productive");
-        transition-property: color, border-color;
-
-        &:hover,
-        &:focus-visible {
-            --icon-background: var(--jkl-color-background-interactive-selected);
-        }
-
-        &:focus-visible {
-            --border-color: transparent;
-            z-index: 1;
-
-            @include jkl.focus-outline($offset: 0);
-        }
-
-        &__arrow {
-            border-radius: 100%;
-            background-color: var(--icon-background);
-            // Justér ikonet i forhold til tekstens baseline
-            translate: 0 17%;
-
-            @include jkl.motion("standard", "productive");
-            transition-property: color, background-color;
-        }
-
-        .jkl-link-list--ordered & {
-            justify-content: flex-start;
-            align-items: unset;
-
-            &__arrow {
-                display: none;
+            &:not(:last-of-type) {
+                .jkl-link-list-link {
+                    border-block-end: var(--border);
+                }
             }
+        }
 
-            &::before {
-                content: counter(link-list, decimal-leading-zero);
-
-                @include jkl.use-font-family("Fremtind Grotesk Mono");
-                color: var(--text-color);
-                font-size: var(--jkl-small-font-size);
+        &:has(ul) {
+            .jkl-link-list-title {
+                margin-inline-start: var(--jkl-unit-05);
             }
         }
     }
