+++ packages/jokul/src/components/link-list/types.ts
@@ -1,18 +1,13 @@
-import type { ComponentProps, FC } from "react";
+import type { HTMLAttributes } from "react";
 import type { PolymorphicPropsWithRef } from "../../utilities/polymorphism/polymorphism.js";
 
-export type UnorderedLinkListProps = React.HTMLAttributes<HTMLUListElement> & {
-    variant: "unordered";
+export type LinkListProps = HTMLAttributes<HTMLUListElement> & {
+    label: string;
+    hideLabel?: boolean;
 };
 
-export type OrderedLinkListProps = React.HTMLAttributes<HTMLOListElement> & {
-    variant: "ordered";
-};
-
-export type LinkListProps = UnorderedLinkListProps | OrderedLinkListProps;
-
-export type LinkComponent = <ElementType extends React.ElementType = "a">(
+export type LinkListLinkComponent = <
+    ElementType extends React.ElementType = "a",
+>(
     props: PolymorphicPropsWithRef<ElementType>,
 ) => React.ReactElement | null;
-
-export type ItemProps = FC<ComponentProps<"li">>;
