+++ packages/jokul/src/styles/styles.scss
@@ -19,7 +19,6 @@
 @use "../components/input-group/styles" as input-group;
 @use "../components/link/styles" as link;
 @use "../components/link-list/styles" as link-list;
-@use "../components-beta/link-list/styles" as link-list-beta;
 @use "../components/list/styles" as list;
 @use "../components/loader/styles" as loader;
 @use "../components/logo/styles" as logo;
